Extract path.c from misc1.c
