+++ src/buffer.c
@@ -55,6 +55,7 @@
 #include "move.h"
 #include "option.h"
 #include "os_unix.h"
+#include "path.h"
 #include "quickfix.h"
 #include "regexp.h"
 #include "screen.h"
