+++ src/file_search.c
@@ -54,6 +54,7 @@
 #include "misc1.h"
 #include "misc2.h"
 #include "os_unix.h"
+#include "path.h"
 #include "tag.h"
 #include "ui.h"
 #include "window.h"
