+++ src/misc2.c
@@ -37,6 +37,7 @@
 #include "option.h"
 #include "ops.h"
 #include "os_unix.h"
+#include "path.h"
 #include "quickfix.h"
 #include "regexp.h"
 #include "screen.h"
