+++ src/ops.c
@@ -35,6 +35,7 @@
 #include "move.h"
 #include "normal.h"
 #include "option.h"
+#include "path.h"
 #include "screen.h"
 #include "search.h"
 #include "term.h"
