+++ src/os_unix.c
@@ -44,6 +44,7 @@
 #include "misc1.h"
 #include "misc2.h"
 #include "garray.h"
+#include "path.h"
 #include "screen.h"
 #include "syntax.h"
 #include "term.h"
