+++ src/screen.c
@@ -113,6 +113,7 @@
 #include "move.h"
 #include "normal.h"
 #include "option.h"
+#include "path.h"
 #include "popupmnu.h"
 #include "quickfix.h"
 #include "regexp.h"
